+++ src/Model/DTO/Admin/ConfigDTO.php
@@ -11,9 +11,9 @@ final class ConfigDTO implements FactorableDTOInterface
         public readonly int $draw_begins_at_day_no,
         public readonly int $draw_begins_at_concrete_day,
         public readonly bool $draw_is_concrete_day_set,
-        public readonly int $lottery_ticket_cost,
-        public readonly int $lottery_required_tickets_sum,
-        public readonly int $fee_basic,
+        public readonly float $lottery_ticket_cost,
+        public readonly float $lottery_required_tickets_sum,
+        public readonly float $fee_basic,
     ) {
     }
 }
